Added import id and import swap functions to fileMng.py and configured interface.py to make use of them.  With import id when selecting a host file creates a temporary file and loads it with the contents of ip.list followed by the contents of the host file selected via an id number entered.
ex:  sudo python3 import 2

This would create a host file using the host file downloaded from url 2
with including ipexceptions and then placing it into /etc .  The temporary
files created are then removed as result

The function of import swap is to store the current host file in place as a
temporary file and then placing a basic host file int /etc .  Doing this
again with out importing a different host file swaps the basic host file
with the previous host file being used.

The point of import swap is to quickly remove ad blocking in the event your
host file is blocking a website you need to make use of and then you can
swap back afterwards.

Content of the basic host file thats made:

127.0.0.1		localhost
::1			localhost