Added scanning function to fileMng.py.  Its applied to scan specifichost files as well as scanning newly downloaded files.
	In the event of IPs being found that are suspicious the downloaded
file will be blacklisted so it can't be used and the user will be told to
remove the file and download link. If bad IPs that don't work are found the
user will be suggested to use another host file but the file will still be
usable as it doesn't contain suspicious IPs.

	Whether or not any IP problems are found the count of them are
printed and ones that are found are logged to files inside the hlist
directory.  These files being SketchLog.idx and BadIpLog.idx "x" being the
correlating host file id number.